requirements.txt
streamlit
deepface
opencv-python
scipy
numpy
Pillow
mtcnn
retina-face
